4.15-4.22

+++ controller/blog.js
@@ -1,27 +1,77 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 // get blogposts
-blogRouter.get('/', async (req, res, next) => {
-    try {
-        const blogs = await Blog.find({});
-        res.json(blogs);
-    } catch (error) {
-        next(error);
-    }
+blogRouter.get('/', async (req, res) => {
+    const blogs = await Blog.find({})
+        .populate('user', {
+            username: 1,
+            name: 1
+        });
+    res.json(blogs);
 });
+
 
 //new blog
 blogRouter.post('/', async (req, res) => {
-    const blog = new Blog(req.body);
-    const savedBlog = await blog.save();
+    const body = req.body;
+
+    const token = req.token;
+
+    if(!token){
+        return res.status(401).json({ error: 'Unauthorized: Token is missing' });
+    }
+
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return res.status(401).json({ error: 'token invalid' });
+    }
+    const user = await User.findById(decodedToken.id);
+
+    const note = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    });
+
+    const savedBlog = await note.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
+
     res.status(201).json(savedBlog);
 });
 
 //delete a specific blog when given the id
 blogRouter.delete('/:id', async (req, res) => {
-    await Blog.findOneAndDelete(req.params.id);
-    res.status(204).end();
+    const token = req.token;
+
+    if(!token){
+        return res.status(401).json({ error: 'Unauthorized: Token is missing' });
+    }
+
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return res.status(401).json({ error: 'token invalid' });
+    }
+
+
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(req.params.id);
+
+    console.log('user ',user);
+    console.log('blog ',blog);
+    if(user._id.toString() === blog.user.toString()){
+        await Blog.findOneAndDelete(req.params.id);
+        user.blogs = user.blogs.filter(blog => blog._id !== req.params.id);
+        await user.save();
+        res.status(204).end();
+    }else{
+        return res.status(401).json({ error: 'not allowed to delete this blog' });
+    }
 });
 
 //get a specific blog when given the id
@@ -46,7 +96,9 @@ blogRouter.put('/:id', async (req, res) => {
         blogs: body.blogs,
     };
 
-    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, { new: true });
+    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, {
+        new: true
+    });
     res.json(updatedBlog);
 });
